interview_module/requirements.txt
torchaudio
omegaconf
torch
fastapi
uvicorn
sounddevice
pydub
scipy
gTTS
openai-whisper
flask
pyaudio
librosa
numpy
python-multipart
pyttsx3
ttsmaker
TeraTTS
ruaccent
rhvoice_wrapper
simpleaudio
transliterate
tensorflow
soundfile
speechkit
python-decouple
lipsync
